Fixed Windows build

+++ code/sptk5/db/BulkQuery.h
@@ -33,7 +33,7 @@ public:
      * @param columnNames       Column names.
      * @param groupSize         Group size.
      */
-    BulkQuery(PoolDatabaseConnection* connection, const String& tableName, const String& serialColumnName, const Strings& columnNames, unsigned groupSize);
+    BulkQuery(PoolDatabaseConnection* connection, const String& tableName, const String& serialColumnName, const Strings& columnNames, size_t groupSize);
 
     /**
      * Insert rows into the table.
@@ -54,20 +54,21 @@ private:
     String                  m_serialColumnName;    ///< Auto increment column name (optional, can be empty).
     Strings                 m_columnNames;         ///< Column names.
     String                  m_tableName;           ///< Table name.
-    unsigned                m_groupSize;           ///< Insert or delete record group size.
+    size_t                  m_groupSize;           ///< Insert or delete record group size.
     PoolDatabaseConnection* m_connection;          ///< Database connection.
     Query                   m_lastInsertedIdQuery; ///< Query that retrieves the last inserted id.
 
-    [[nodiscard]] static String makeInsertSQL(DatabaseConnectionType connectionType, const String& tableName, const String& keyColumnName, const Strings& columnNames, unsigned groupSize);
-    [[nodiscard]] static String makeOracleInsertSQL(const String& tableName, const Strings& columnNames, unsigned groupSize);
-    [[nodiscard]] static String makeGenericInsertSQL(const String& tableName, const Strings& columnNames, unsigned groupSize, const String& intoAttribute = "");
-    [[nodiscard]] static String makeSqlite3InsertSQL(const String& tableName, const Strings& columnNames, unsigned groupSize);
-    [[nodiscard]] static String makeGenericDeleteSQL(const String& tableName, const String& keyColumnName, unsigned int groupSize);
+    [[nodiscard]] static String makeInsertSQL(DatabaseConnectionType connectionType, const String& tableName, const String& keyColumnName, const Strings& columnNames, size_t groupSize);
+    [[nodiscard]] static String makeOracleInsertSQL(const String& tableName, const Strings& columnNames, size_t groupSize);
+    [[nodiscard]] static String makeGenericInsertSQL(const String& tableName, const Strings& columnNames, size_t groupSize, const String& intoAttribute = "");
+    [[nodiscard]] static String makeSqlite3InsertSQL(const String& tableName, const Strings& columnNames, size_t groupSize);
+    [[nodiscard]] static String makeGenericDeleteSQL(const String& tableName, const String& keyColumnName, size_t groupSize);
 
     void        beginInsert(bool& startedTransaction) const;
     void        commitInsert() const;
     bool        reserveInsertIds(const String& tableName, const std::vector<std::vector<Variant>>& rows, std::vector<int64_t>& insertedIds);
-    size_t      insertGroupRows(Query& insertQuery, std::vector<VariantVector>::const_iterator startRow, std::vector<VariantVector>::const_iterator end, std::vector<long>& insertedIds, bool useReservedIds, size_t serialColumnIndex, size_t& reservedIdOffset);
+    size_t      insertGroupRows(Query& insertQuery, std::vector<VariantVector>::const_iterator startRow, std::vector<VariantVector>::const_iterator end, 
+								std::vector<int64_t>& insertedIds, bool useReservedIds, size_t serialColumnIndex, size_t& reservedIdOffset);
     static void deleteGroupRows(Query& insertQuery, VariantVector::const_iterator startKey, VariantVector::const_iterator end);
 };
